+++ api/src/main/java/org/grnet/status/api/endpoints/AutomationEndpoint.java
@@ -3,6 +3,7 @@ package org.grnet.status.api.endpoints;
 import io.quarkus.security.Authenticated;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
@@ -14,18 +15,23 @@ import org.eclipse.microprofile.openapi.annotations.media.Content;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
+import org.eclipse.microprofile.openapi.annotations.security.SecurityRequirement;
 import org.eclipse.microprofile.openapi.annotations.security.SecurityScheme;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.grnet.endpoint.scanner.runtime.ParamRef;
 import org.grnet.endpoint.scanner.runtime.ParamType;
 import org.grnet.endpoint.scanner.runtime.SecuredEndpoint;
+import org.grnet.status.api.resolvers.CheckDateFormat;
 import org.grnet.status.constraints.NotFoundEntity;
 import org.grnet.status.dtos.InformativeResponse;
+import org.grnet.status.dtos.downtime.DailyDowntimeResponse;
 import org.grnet.status.dtos.tenant.status.TenantStatusDto;
 import org.grnet.status.dtos.tenant.status.TenantStatusFullResponse;
+import org.grnet.status.entities.Downtime;
 import org.grnet.status.enums.resources.InvitationResource;
 import org.grnet.status.enums.resources.TenantResource;
 import org.grnet.status.repositories.TenantRepository;
+import org.grnet.status.services.DowntimeService;
 import org.grnet.status.services.TenantService;
 
 
@@ -41,6 +47,8 @@ public class AutomationEndpoint {
 
     @Inject
     TenantService tenantService;
+    @Inject
+    DowntimeService downtimeService;
 
     @Tag(name = "Automation")
     @Operation(
@@ -151,7 +159,8 @@ public class AutomationEndpoint {
                             referTo = TenantResource.class
                     )
             }
-    )    public Response getTenantStatus(
+    )
+    public Response getTenantStatus(
             @Parameter(
                     description = "The ID of the tenant to update status.",
                     required = true,
@@ -164,5 +173,81 @@ public class AutomationEndpoint {
         var status = tenantService.getTenantStatus(id);
 
         return Response.ok().entity(status).build();
+    }
+
+
+    @Tag(name = "Automation")
+    @Operation(
+            summary = "Fetch a daily downtime for a tenant.",
+            description = "Returns the tenant's specific daily downtime"
+    )
+    @APIResponse(
+            responseCode = "200",
+            description = "Downtimes fetched successfully.",
+            content = @Content(schema = @Schema(implementation = DailyDowntimeResponse.class))
+    )
+    @APIResponse(
+            responseCode = "400",
+            description = "Invalid request payload.",
+            content = @Content(schema = @Schema(implementation = InformativeResponse.class))
+    )
+    @APIResponse(
+            responseCode = "401",
+            description = "User not authenticated.",
+            content = @Content(schema = @Schema(implementation = InformativeResponse.class))
+    )
+    @APIResponse(
+            responseCode = "403",
+            description = "Not permitted.",
+            content = @Content(schema = @Schema(implementation = InformativeResponse.class))
+    )
+    @APIResponse(
+            responseCode = "404",
+            description = "Downtimes not found.",
+            content = @Content(schema = @Schema(implementation = InformativeResponse.class))
+    )
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal server error.",
+            content = @Content(schema = @Schema(implementation = InformativeResponse.class))
+    )
+
+    @GET
+    @Path("/tenants/{id}/downtimes/daily")
+    @Produces(MediaType.APPLICATION_JSON)
+    @SecuredEndpoint(
+            params = {
+                    @ParamRef(
+                            param = "id",
+                            type = ParamType.PATH,
+                            referTo = TenantResource.class
+                    )
+            }
+    )
+    public Response getDailyDowntimes(
+            @Parameter(
+                    description = "The ID of the tenant to fetch downtimes.",
+                    required = true,
+                    example = "42c1152d-e23c-4a19-b51a-b27f1eb7f37f",
+                    schema = @Schema(type = SchemaType.STRING)) @PathParam("id")
+            @Valid @NotFoundEntity(repository = TenantRepository.class, message = "There is no Tenant with the following id: ") String id,
+            @Parameter(
+                    name = "date",
+                    description = "UTC date in yyyy-MM-dd format",
+                    example = "2026-07-22",
+                    required = true
+            )
+            @CheckDateFormat(
+                    pattern = "yyyy-MM-dd",
+                    message = "Valid date format is yyyy-MM-dd."
+            )
+            @QueryParam("date")
+            @NotNull
+            String date
+    ) {
+
+        return Response.ok(
+                downtimeService.fetchDailyDowntimes(id, date)
+        ).build();
     }
 }